+++ PageReindex.cpp
@@ -61,7 +61,7 @@ bool sendPageReindex ( TcpSocket *s , HttpRequest *r ) {
 	catch ( ... ) {
 		g_errno = ENOMEM;
 		log("PageTagdb: new(%i): %s", 
-		    sizeof(State13),mstrerror(g_errno));
+		    (int)sizeof(State13),mstrerror(g_errno));
 		return g_httpServer.sendErrorReply(s,500,mstrerror(g_errno));}
 	mnew ( st , sizeof(State13) , "PageReindex" );
